Split slash_commands into module
